voice/requirements.txt
openwakeword
pvporcupine
sounddevice
numpy
scipy
httpx
python-dotenv
paho-mqtt
prometheus-client
